Move funding communication actions that are communicated to communication module
