+++ packages/wallet/src/redux/protocols/funding/player-b/actions.ts
@@ -1,6 +1,8 @@
 import { BaseProcessAction } from '../../actions';
 import { PlayerIndex } from '../../../types';
 import { Strategy } from '..';
+import { strategyProposed, StrategyProposed, STRATEGY_PROPOSED } from '../../../../communication';
+export { strategyProposed, StrategyProposed, STRATEGY_PROPOSED };
 
 export type FundingAction =
   | StrategyProposed
@@ -9,17 +11,11 @@ export type FundingAction =
   | StrategyRejected
   | Cancelled;
 
-export const STRATEGY_PROPOSED = 'WALLET.FUNDING.STRATEGY_PROPOSED';
 export const STRATEGY_APPROVED = 'WALLET.FUNDING.STRATEGY_APPROVED';
 export const FUNDING_SUCCESS_ACKNOWLEDGED = 'WALLET.FUNDING.FUNDING_SUCCESS_ACKNOWLEDGED';
 export const STRATEGY_REJECTED = 'WALLET.FUNDING.STRATEGY_REJECTED';
 export const CANCELLED = 'WALLET.FUNDING.CANCELLED';
 export const CANCELLED_BY_OPPONENT = 'WALLET.FUNDING.CANCELLED_BY_OPPONENT';
-
-export interface StrategyProposed extends BaseProcessAction {
-  type: typeof STRATEGY_PROPOSED;
-  strategy: Strategy;
-}
 
 export interface StrategyApproved extends BaseProcessAction {
   type: typeof STRATEGY_APPROVED;
@@ -42,12 +38,6 @@ export interface Cancelled extends BaseProcessAction {
 // --------
 // Creators
 // --------
-
-export const strategyProposed = (processId: string, strategy): StrategyProposed => ({
-  type: STRATEGY_PROPOSED,
-  processId,
-  strategy,
-});
 
 export const strategyApproved = (processId: string, strategy): StrategyApproved => ({
   type: STRATEGY_APPROVED,
